pyproject.toml
[project]
name = "pennywise"
version = "0.1.0"
description = ""
authors = [
    {name = "plmeister",email = "plmeister@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

